requirements.txt
flask ==3.0.2
mysql-connector-python == 8.3.0